Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    double float a;
+    float a;
     scanf("%0.4f",&a);
     printf("You entered: %0.4f",a);
     return 0;
